refactor: split topic resume and login choice flows
